Refactor: ProductService DTO로 반환
