+++ src/main/java/kitchenpos/product/tobe/application/dto/response/ProductResponse.java
@@ -1,0 +1,7 @@
+package kitchenpos.product.tobe.application.dto.response;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public record ProductResponse(UUID id, String name, BigDecimal price) {
+}
